requirements.txt
flask==3.1.0
pdfplumber==0.11.5
pytesseract==0.3.13
Pillow==11.1.0
spacy==3.8.4
werkzeug==3.1.3
gunicorn==23.0.0
pytest==8.3.5
